improve feat highlighting
