+++ packages/widgetbook/lib/next.dart
@@ -4,6 +4,7 @@ library widgetbook.next;
 import 'package:meta/meta.dart';
 
 export 'src/next/component_metadata.dart';
+export 'src/next/next_feat_wrapper.dart';
 export 'src/next/widgetbook_args.dart';
 export 'src/next/widgetbook_component.dart';
 export 'src/next/widgetbook_mode.dart';
